+++ timetable_bot/config/default.py
@@ -17,6 +17,8 @@ class DefaultSettings(BaseSettings):
     DB_CONNECT_RETRY: int = environ.get("DB_CONNECT_RETRY", 20)
     DB_POOL_SIZE: int = environ.get("DB_POOL_SIZE", 15)
 
+    TIMEZONE_OFFSET: int = 3
+
     @property
     def database_settings(self) -> dict:
         """
