--- timetable_bot/db/migrator/versions/2022_09_01_2119-11f80ab59a15_.py
@@ -1,39 +1,0 @@
-"""empty message
-
-Revision ID: 11f80ab59a15
-Revises: 
-Create Date: 2022-09-01 21:19:19.470236
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision = '11f80ab59a15'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('user',
-    sa.Column('id', postgresql.UUID(as_uuid=True), server_default=sa.text('gen_random_uuid()'), nullable=False),
-    sa.Column('dt_created', postgresql.TIMESTAMP(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
-    sa.Column('tg_id', sa.INTEGER(), nullable=False),
-    sa.Column('username', sa.TEXT(), nullable=False),
-    sa.Column('group', sa.TEXT(), nullable=False),
-    sa.PrimaryKeyConstraint('id', name=op.f('pk__user')),
-    sa.UniqueConstraint('id', name=op.f('uq__user__id')),
-    sa.UniqueConstraint('username', name=op.f('uq__user__username'))
-    )
-    op.create_index(op.f('ix__user__tg_id'), 'user', ['tg_id'], unique=True)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix__user__tg_id'), table_name='user')
-    op.drop_table('user')
-    # ### end Alembic commands ###
