--- timetable_bot/db/migrator/versions/2022_09_02_1430-53a26b4a4520_.py
@@ -1,28 +1,0 @@
-"""empty message
-
-Revision ID: 53a26b4a4520
-Revises: 4214cd7f6d87
-Create Date: 2022-09-02 14:30:34.217694
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '53a26b4a4520'
-down_revision = '4214cd7f6d87'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_unique_constraint(op.f('uq__user__id'), 'user', ['id'])
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(op.f('uq__user__id'), 'user', type_='unique')
-    # ### end Alembic commands ###
