+++ timetable_bot/handlers/callback.py
@@ -21,7 +21,6 @@ async def handle_day_select(call: types.CallbackQuery):
     group = await utils.get_user_group(call.from_user.id)
     if not group:
         await call.message.answer("надо выбрать группу:", reply_markup=kb.group_sel_kb)
-        await call.answer()
     else:
         day = utils.parse_sel_day_data(call.data)
         result = await utils.get_day(group, day)
@@ -30,7 +29,7 @@ async def handle_day_select(call: types.CallbackQuery):
             await call.message.edit_text(result, reply_markup=kb.day_switch_kb(day_for_button))
         except:
             logging.warning("same text, didn't edit")
-        await call.answer()
+    await call.answer()
 
 
 async def handle_day_switch(call: types.CallbackQuery):
@@ -40,19 +39,18 @@ async def handle_day_switch(call: types.CallbackQuery):
     group = await utils.get_user_group(call.from_user.id)
     if not group:
         await call.message.answer("надо выбрать группу:", reply_markup=kb.group_sel_kb)
-        await call.answer()
     else:
-        new_day = utils.parse_day_switch_data(call.data)
-        if new_day == "menu":
-            try:
-                await call.message.edit_text("выбери день.\n", reply_markup=kb.day_sel_kb)
-            except:
-                logging.warning("same text, didn't edit")
-            await call.answer()
+        button_pressed = utils.parse_day_switch_data(call.data)
+        if button_pressed == "menu":
+            reply_kb = kb.day_sel_kb
+            msg = "выбери день.\n"
         else:
-            result = await utils.get_day(group, utils.weeknum_to_weekday(int(new_day)))
-            try:
-                await call.message.edit_text(result, reply_markup=kb.day_switch_kb(int(new_day)))
-            except:
-                logging.warning("same text, didn't edit")
-            await call.answer()
+            new_day = int(button_pressed)
+            reply_kb = kb.day_switch_kb(new_day)
+            msg = await utils.get_day(group, utils.weeknum_to_weekday(new_day))
+
+        try:
+            await call.message.edit_text(msg, reply_markup=reply_kb)
+        except:
+            logging.warning("same text, didn't edit")
+    await call.answer()
