+++ timetable_bot/handlers/commands.py
@@ -3,9 +3,10 @@ from datetime import timedelta
 
 import timetable_bot.keyboards as kb
 import timetable_bot.utils as utils
+from timetable_bot.config import DefaultSettings
 
 
-TD = timedelta(hours=3)
+TD = timedelta(hours=DefaultSettings().TIMEZONE_OFFSET)
 
 async def send_hello(message: types.Message):
     """
@@ -32,7 +33,7 @@ async def send_echo(message: types.Message):
             await send_week_schedule(message)
         case _:
             await message.reply(message.from_user.first_name + " говорит: " + message.text + "\nвыбери день:", reply_markup=kb.day_sel_kb)
-    
+
 
 async def send_week_schedule(message: types.Message):
     """
@@ -94,57 +95,3 @@ async def del_me_from_db(message: types.Message):
     """
     result = await utils.del_user_from_db(message.from_user.id)
     await message.reply(result)
-
-
-# # callback handlers #
-# async def handle_user_group(call: types.CallbackQuery):
-#     """
-#     Отвечаем на нажатие инлайн клавы с выбором группы (меняем группу)
-#     """
-#     res = await utils.set_user_group(call.from_user, call.data)
-#     await call.message.answer(res, reply_markup=kb.smile_kb)
-#     await call.answer()
-
-
-# async def handle_day_select(call: types.CallbackQuery):
-#     """
-#     Отвечаем на нажатие инлайн клавы с выбором дня недели
-#     """
-#     group = await utils.get_user_group(call.from_user.id)
-#     if not group:
-#         await call.message.answer("надо выбрать группу:", reply_markup=kb.group_sel_kb)
-#         await call.answer()
-#     else:
-#         day = utils.parse_sel_day_data(call.data)
-#         result = await utils.get_day(group, day)
-#         day_for_button = utils.weekday_to_weeknum(day)
-#         try:
-#             await call.message.edit_text(result, reply_markup=kb.day_switch_kb(day_for_button))
-#         except:
-#             logging.warning("same text, didn't edit")
-#         await call.answer()
-
-
-# async def handle_day_switch(call: types.CallbackQuery):
-#     """
-#     Отвечаем на нажатие инлайн клавы с переключением дня недели
-#     """
-#     group = await utils.get_user_group(call.from_user.id)
-#     if not group:
-#         await call.message.answer("надо выбрать группу:", reply_markup=kb.group_sel_kb)
-#         await call.answer()
-#     else:
-#         new_day = utils.parse_day_switch_data(call.data)
-#         if new_day == "menu":
-#             try:
-#                 await call.message.edit_text("выбери день.\n", reply_markup=kb.day_sel_kb)
-#             except:
-#                 logging.warning("same text, didn't edit")
-#             await call.answer()
-#         else:
-#             result = await utils.get_day(group, utils.weeknum_to_weekday(int(new_day)))
-#             try:
-#                 await call.message.edit_text(result, reply_markup=kb.day_switch_kb(int(new_day)))
-#             except:
-#                 logging.warning("same text, didn't edit")
-#             await call.answer()
