+++ timetable_bot/keyboards/__init__.py
@@ -1,4 +1,5 @@
-from .reply import group_sel_kb, group_callback_data, day_sel_kb, nothing, smile_kb, day_switch_kb
+from .inline import group_sel_kb, group_callback_data, day_sel_kb, day_switch_kb
+from .reply import smile_kb, nothing
 
 
 __all__ = [
